New pyposit class with init
